restructure the building-hours detail views
