--- source/views/building-hours/detail.ios.js
@@ -1,202 +1,0 @@
-// @flow
-
-import React from 'react'
-import {View, Text, StyleSheet} from 'react-native'
-import {buildingImages} from '../../../images/building-images'
-import type {
-  BuildingType,
-  DayOfWeekEnumType,
-  SingleBuildingScheduleType,
-} from './types'
-import type momentT from 'moment'
-import moment from 'moment-timezone'
-import {TableView, Section, Cell} from 'react-native-tableview-simple'
-import ParallaxView from 'react-native-parallax-view'
-import * as c from '../components/colors'
-import {
-  normalizeBuildingSchedule,
-  formatBuildingTimes,
-  summarizeDays,
-  getShortBuildingStatus,
-  isBuildingOpenAtMoment,
-} from './building-hours-helpers'
-
-const CENTRAL_TZ = 'America/Winnipeg'
-const transparentPixel = require('../../../images/transparent.png')
-
-const styles = StyleSheet.create({
-  title: {
-    paddingTop: 16,
-    paddingBottom: 4,
-    paddingHorizontal: 8,
-  },
-  name: {
-    textAlign: 'center',
-    color: c.black,
-    fontSize: 32,
-    fontWeight: '300',
-  },
-  subtitle: {
-    paddingTop: 8,
-    paddingHorizontal: 8,
-  },
-  subtitleText: {
-    fontSize: 18,
-  },
-  badge: {
-    marginTop: 20,
-    paddingHorizontal: 12,
-    paddingVertical: 4,
-    borderRadius: 5,
-    alignSelf: 'center',
-  },
-  badgeText: {
-    color: c.white,
-    fontSize: 18,
-  },
-  scrollableStyle: {
-    backgroundColor: c.iosLightBackground,
-  },
-  bold: {
-    fontWeight: '600',
-  },
-})
-
-export class BuildingHoursDetailView extends React.Component {
-  static navigationOptions = ({navigation}) => {
-    return {
-      title: navigation.state.params.building.name,
-    }
-  }
-
-  state: {intervalId: number, now: momentT} = {
-    intervalId: 0,
-    // now: moment.tz('Wed 7:25pm', 'ddd h:mma', null, CENTRAL_TZ),
-    now: moment.tz(CENTRAL_TZ),
-  }
-
-  componentWillMount() {
-    // This updates the screen every ten seconds, so that the building
-    // info statuses are updated without needing to leave and come back.
-    this.setState({intervalId: setInterval(this.updateTime, 10000)})
-  }
-
-  componentWillUnmount() {
-    clearTimeout(this.state.intervalId)
-  }
-
-  props: {navigation: {state: {params: {building: BuildingType}}}}
-
-  updateTime = () => {
-    this.setState({now: moment.tz(CENTRAL_TZ)})
-  }
-
-  render() {
-    const info = this.props.navigation.state.params.building
-    const {now} = this.state
-
-    const headerImage = info.image
-      ? buildingImages[info.image]
-      : transparentPixel
-    const openStatus = getShortBuildingStatus(info, this.state.now)
-    const schedules = normalizeBuildingSchedule(info, this.state.now)
-    const dayOfWeek = ((this.state.now.format('dd'): any): DayOfWeekEnumType)
-
-    return (
-      <ParallaxView
-        backgroundSource={headerImage}
-        windowHeight={100}
-        scrollableViewStyle={styles.scrollableStyle}
-      >
-        <View>
-          <Header building={info} />
-
-          <Badge status={openStatus} />
-
-          <TableView>
-            {schedules.map(set =>
-              <Section
-                key={set.title}
-                header={set.title.toUpperCase()}
-                footer={set.notes}
-              >
-                {set.hours.map((schedule, i) =>
-                  <ScheduleRow
-                    key={i}
-                    now={now}
-                    schedule={schedule}
-                    isActive={
-                      set.isPhysicallyOpen !== false &&
-                      schedule.days.includes(dayOfWeek) &&
-                      isBuildingOpenAtMoment(schedule, this.state.now)
-                    }
-                  />,
-                )}
-              </Section>,
-            )}
-          </TableView>
-        </View>
-      </ParallaxView>
-    )
-  }
-}
-
-function Header({building}) {
-  const abbr = building.abbreviation
-    ? <Text style={styles.abbr}> ({building.abbreviation})</Text>
-    : null
-
-  const title = (
-    <Text selectable={true} style={styles.name}>{building.name}{abbr}</Text>
-  )
-
-  const subtitle = building.subtitle
-    ? <View style={styles.subtitle}>
-        <Text selectable={true} style={[styles.name, styles.subtitleText]}>
-          {building.subtitle}
-        </Text>
-      </View>
-    : null
-
-  return (
-    <View>
-      <View style={styles.title}>{title}</View>
-      {subtitle}
-    </View>
-  )
-}
-
-function Badge({status}) {
-  const bgColors = {
-    Open: c.moneyGreen,
-    Closed: c.salmon,
-  }
-
-  return (
-    <View
-      style={[styles.badge, {backgroundColor: bgColors[status] || c.goldenrod}]}
-    >
-      <Text selectable={true} style={styles.badgeText}>{status}</Text>
-    </View>
-  )
-}
-
-function ScheduleRow({
-  schedule,
-  isActive,
-  now,
-}: {
-  schedule: SingleBuildingScheduleType,
-  isActive: boolean,
-  now: momentT,
-}) {
-  return (
-    <Cell
-      cellStyle="RightDetail"
-      title={summarizeDays(schedule.days)}
-      titleTextStyle={isActive ? styles.bold : null}
-      detail={formatBuildingTimes(schedule, now)}
-      detailTextStyle={isActive ? styles.bold : null}
-    />
-  )
-}
