+++ source/views/building-hours/detail/schedule-row.android.js
@@ -1,0 +1,57 @@
+/**
+ * @flow
+ *
+ * <ScheduleRow/> renders a single row of the schedule information.
+ */
+
+import React from 'react'
+import {View, Text, StyleSheet} from 'react-native'
+import type {SingleBuildingScheduleType} from '../types'
+
+import {formatBuildingTimes, summarizeDays} from '../building-hours-helpers'
+
+export class ScheduleRow extends React.PureComponent {
+  props: {
+    schedule: SingleBuildingScheduleType,
+    isActive: boolean,
+    now: moment,
+  }
+
+  render() {
+    const {schedule, isActive, now} = this.props
+    return (
+      <View style={styles.scheduleRow}>
+        <StyledText style={[styles.scheduleDays, isActive && styles.bold]}>
+          {summarizeDays(schedule.days)}
+        </StyledText>
+
+        <StyledText style={[styles.scheduleHours, isActive && styles.bold]}>
+          {formatBuildingTimes(schedule, now)}
+        </StyledText>
+      </View>
+    )
+  }
+}
+
+const StyledText = ({children, style}) =>
+  <Text selectable={true} numberOfLines={1} style={style}>
+    {children}
+  </Text>
+
+const styles = StyleSheet.create({
+  bold: {
+    fontWeight: 'bold',
+  },
+  scheduleRow: {
+    flexDirection: 'row',
+    paddingVertical: 6,
+  },
+  scheduleDays: {
+    flex: 1,
+    textAlign: 'right',
+    paddingRight: 16,
+  },
+  scheduleHours: {
+    flex: 2,
+  },
+})
